Give up on supporting Pandoc v1
The HTML generated by Pandoc v2 for syntax-highlighted code blocks is
significantly different from v1 (it uses :before pseudo selectors, but
before it was tables). It's too hard to support both representations
with the same CSS, so we're dropping support for v1.

If you need v1 support, consider starting from a revision before this
commit.